:building_construction: Split up controller & service for pages and landingPages
